Added Results Export File for 1-nn. I'll be adding a K-NN option in the next update..

+++ src/main/java/DataRow.java
@@ -11,8 +11,8 @@ public class DataRow implements Serializable {
     int id ;
     public DataRow(){
         classLabel = -1 ;
-        columns = new ArrayList<Integer>();
+        columns = new ArrayList<Double>();
     }
-    public List<Integer> columns ;
+    public List<Double> columns ;
     public int classLabel;
 }
